Title: 1까지 나누기, Time: 9ms, Memory: 1MB, Status: Passed - Codetree

+++ 260121/1까지 나누기/divide-by-1.go
@@ -3,12 +3,12 @@ package main
 import "fmt"
 
 func main() {
-	var n float64
-	fmt.Scanf("%f", &n)
+	var n int
+	fmt.Scanf("%d", &n)
 	i := 1
 	cnt := 0
 	for {
-		n /= float64(i)
+		n /= i
 		cnt++
 		if n <= 1 {
 			break
